requirements.txt
Flask
Flask-sqlalchemy
ujson
mysql-python
itsdangerous
passlib
tornado
dnspython